src/main/java/org/example/thirdlab/products/Product.java
package org.example.thirdlab.products;

import java.util.Objects;

public class Product {
    private String name;
    private String description;

    public Product(String name, String description) {
        if (name.isBlank()) {
            //isBlank возвращает true в любом случае "пустого" имени
            throw new IllegalArgumentException("Некорректное имя товара");
        }
        this.name = name;
        this.description = description;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (!(o instanceof Product product)) return false;
        return Objects.equals(getName(), product.getName()) && Objects.equals(getDescription(), product.getDescription());
    }

    @Override
    public int hashCode() {
        return Objects.hash(getName(), getDescription());
    }

    @Override
    public String toString() {
        return "Product{" +
                "name='" + name + '\'' +
                ", description='" + description + '\'' +
                '}';
    }
}
